Add facade top layer service to help co-ordinate the use of prisoner contact service for contact retrieval logic and restrictions service for handling restriction logic. Used as an orchestration layer.
